refactor: move NotificationTransportConfig to homeboy-extension-contract
NotificationTransportConfig (+ NOTIFICATION_TRANSPORT_SCHEMA const, default fn,
and validate() impl) is a pure serde config type declaring an extension-owned
notification command. It depends only on serde + homeboy_error (both already
contract deps), so move it from homeboy-core's extension::manifest into
homeboy-extension-contract and re-export from extension::manifest to keep the
existing crate::extension::{NotificationTransportConfig, NOTIFICATION_TRANSPORT_
SCHEMA} call sites unchanged.

Continues the extension-contract keystone campaign: draining pure config types
out of the extension subsystem so a future homeboy-extension extraction depends
on a slim seam. Maintainability win on its own — core gets more modular.

Refs #8425
